# openapi.yaml
openapi: 3.1.0
info:
  title: DVI API Bridge
  version: 1.0.0
servers:
  - url: https://dvi.tigertown.net
paths:
  /dvi/start:
    post:
      operationId: startInspection
      summary: Move inspection to "Start" status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ro_number]
              properties:
                ro_number:
                  type: string
      responses:
        '200':
          description: Status updated successfully.
...
  /dvi/pma_complete:
    post:
      operationId: completePma
      summary: Complete PMA process for a repair order.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ro_number]
              properties:
                ro_number:
                  type: string
      responses:
        '200':
          description: PMA process completed.
...
  /dvi/iso_inspection:
    post:
      operationId: isoInspection
      summary: Perform ISO inspection with comments and optional images.
      description: >
        Saves ISO comments via the official R.O.Writer WebForm, attaches image blobs,
        and optionally moves the inspection status to "Start" then "Complete".
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ro_number, rowid]
              properties:
                ro_number:
                  type: string
                comments:
                  type: string
                image_paths:
                  type: array
                  items:
                    type: string
                rowid:
                  type: string
                move_to_start:
                  type: boolean
                move_to_complete:
                  type: boolean
      responses:
        '200':
          description: ISO inspection processed successfully.
...